Trait forest_bitfield::iter::RangeIterator

pub trait RangeIterator: Iterator<Item = Range<usize>> + Sized {
    fn union<R: RangeIterator>(self, other: R) -> Combine<Self, R, Union> { ... }

    fn intersection<R: RangeIterator>(
        self, 
        other: R
    ) -> Combine<Self, R, Intersection> { ... }

    fn difference<R: RangeIterator>(
        self, 
        other: R
    ) -> Combine<Self, R, Difference> { ... }

    fn symmetric_difference<R: RangeIterator>(
        self, 
        other: R
    ) -> Combine<Self, R, SymmetricDifference> { ... }

    fn cut<R: RangeIterator>(self, other: R) -> Combine<Self, R, Cut> { ... }

    fn skip_bits(self, n: usize) -> Skip<Self> { ... }

    fn take_bits(self, n: usize) -> Take<Self> { ... }
}

A trait for iterators over Range<usize>.

Requirements:

all ranges are non-empty

the ranges are in ascending order

no two ranges overlap or touch

Provided methods

fn union<R: RangeIterator>(self, other: R) -> Combine<Self, R, Union>

Returns a new RangeIterator over the bits that are in self, in other, or in both.

fn intersection<R: RangeIterator>(
self,
other: R
) -> Combine<Self, R, Intersection>

Returns a new RangeIterator over the bits that are in both self and other.

fn difference<R: RangeIterator>(self, other: R) -> Combine<Self, R, Difference>

Returns a new RangeIterator over the bits that are in self but not in other.

fn symmetric_difference<R: RangeIterator>(
self,
other: R
) -> Combine<Self, R, SymmetricDifference>

Returns a new RangeIterator over the bits that are in self or in other, but not in both.

fn cut<R: RangeIterator>(self, other: R) -> Combine<Self, R, Cut>

Returns a new RangeIterator over the bits in self that remain after “cutting” out the bits in other, and shifting remaining bits to the left if necessary. For example:

lhs:     xx-xxx--x
rhs:     -xx-x----

cut:     x  x x--x
output:  xxx--x

fn skip_bits(self, n: usize) -> Skip<Self>

Returns a new RangeIterator over the bits in self after skipping the first n bits.

fn take_bits(self, n: usize) -> Take<Self>

Returns a new RangeIterator over the first n bits in self.

Loading content...