Struct forest_bitfield::BitField

pub struct BitField { /* fields omitted */ }

A bit field with buffered insertion/removal that serializes to/from RLE+. Similar to HashSet<usize>, but more memory-efficient when long runs of 1s and 0s are present.

Implementations

impl BitField

pub fn from_bytes(bytes: &[u8]) -> Result<Self, &'static str>

Decodes RLE+ encoded bytes into a bit field.

pub fn to_bytes(&self) -> Vec<u8>

Turns a bit field into its RLE+ encoded form.

impl BitField

pub fn new() -> Self

Creates an empty bit field.

pub fn from_ranges(iter: impl RangeIterator) -> Self

Creates a new bit field from a RangeIterator.

pub fn set(&mut self, bit: usize)

Adds the bit at a given index to the bit field.

pub fn unset(&mut self, bit: usize)

Removes the bit at a given index from the bit field.

pub fn get(&self, index: usize) -> bool

Returns true if the bit field contains the bit at a given index.

pub fn first(&self) -> Option<usize>

Returns the index of the lowest bit present in the bit field.

pub fn iter(&self) -> impl Iterator<Item = usize> + '_

Returns an iterator over the indices of the bit field’s set bits.

pub fn bounded_iter(
&self,
max: usize
) -> Result<impl Iterator<Item = usize> + '_, &'static str>

Returns an iterator over the indices of the bit field’s set bits if the number of set bits in the bit field does not exceed max. Returns an error otherwise.

pub fn ranges(&self) -> impl RangeIterator + '_

Returns an iterator over the ranges of set bits that make up the bit field. The ranges are in ascending order, are non-empty, and don’t overlap.

pub fn is_empty(&self) -> bool

Returns true if the bit field is empty.

pub fn slice(&self, start: usize, len: usize) -> Result<Self, &'static str>

Returns a slice of the bit field with the start index of set bits and number of bits to include in the slice. Returns an error if the bit field contains fewer than start + len set bits.

pub fn len(&self) -> usize

Returns the number of set bits in the bit field.

pub fn cut(&self, other: &Self) -> Self

Returns a new bit field containing the bits in self that remain after “cutting” out the bits in other, and shifting remaining bits to the left if necessary. For example:

lhs:     xx-xxx--x
rhs:     -xx-x----

cut:     x  x x--x
output:  xxx--x

pub fn union<'a>(bitfields: impl IntoIterator<Item = &'a Self>) -> Self

Returns the union of the given bit fields as a new bit field.

pub fn contains_any(&self, other: &BitField) -> bool

Returns true if self overlaps with other.

pub fn contains_all(&self, other: &BitField) -> bool

Returns true if the self is a superset of other.